Move coden tests to modules
This commit moves a couple older codegen tests into modules.
